refactor: extract forge installer logic into forgeinstall subpackage
